requirements.txt
Flask==2.2.2
google-cloud-datastore==2.8.2
python-jose
flask-cors
python-dotenv
requests
authlib
protobuf==3.20.*